stream: move legacy interfaces to stream.d.ts
